requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
docutils==0.16
idna==2.9
Kivy==1.11.1
Kivy-Garden==0.1.4
Pygments==2.6.1
requests==2.23.0
urllib3==1.25.8
python-docx==0.8.10
pyinstaller==3.6
